+++ src/getType.ts
@@ -1,4 +1,4 @@
-import { Schema } from 'mongoose'
+import { Types } from 'mongoose'
 
 import getRealType from "./getRealType"
 
@@ -6,18 +6,18 @@ const getType = (graphqlType: any) => {
   const type = getRealType(graphqlType)
   switch (type) {
     case 'Int': return Number
-    case 'Float': return Number
+    case 'Float': return Types.Decimal128
     case 'Boolean': return Boolean
     case 'String': return String
-    case 'ID': return Schema.Types.ObjectId
+    case 'ID': return Types.ObjectId
     case 'Date': return Date
 
     case '[Int]': return [Number]
-    case '[Float]': return [Number]
+    case '[Float]': return [Types.Decimal128]
     case '[Boolean]': return [Boolean]
     case '[String]': return [String]
     case '[Date]': return [Date]
-    case '[ID]': return [Schema.Types.ObjectId]
+    case '[ID]': return [Types.ObjectId]
   }
 }